@@ -1,11 +1,11 @@
 import { Component } from '@angular/core';
 
 @Component({
-  selector: 'app-corp-identity',
-  standalone: true,
-  imports: [],
-  templateUrl: './corp-identity.component.html',
-  styleUrl: './corp-identity.component.css'
+    standalone: true,
+    selector: 'app-corp-identity',
+    imports: [],
+    templateUrl: './corp-identity.component.html',
+    styleUrl: './corp-identity.component.css'
 })
 export class CorpIdentityComponent {
 
