@@ -7,11 +7,11 @@ import { Access } from '@interfaces/user';
 import { NotificationService } from '@services/notification.service';
 
 @Component({
-  selector: 'app-login',
-  standalone: true,
-  imports: [],
-  templateUrl: './login.component.html',
-  styleUrl: './login.component.css'
+    standalone: true,
+    selector: 'app-login',
+    imports: [],
+    templateUrl: './login.component.html',
+    styleUrl: './login.component.css'
 })
 export class LoginComponent {
 
