@@ -1,11 +1,11 @@
 import { Component } from '@angular/core';
 
 @Component({
-  selector: 'app-no-validated',
-  standalone: true,
-  imports: [],
-  templateUrl: './no-validated.component.html',
-  styleUrl: './no-validated.component.css'
+    standalone: true,
+    selector: 'app-no-validated',
+    imports: [],
+    templateUrl: './no-validated.component.html',
+    styleUrl: './no-validated.component.css'
 })
 export class NoValidatedComponent {
 
