@@ -1,11 +1,11 @@
 import { Component } from '@angular/core';
 
 @Component({
-  selector: 'app-privacy-policy',
-  standalone: true,
-  imports: [],
-  templateUrl: './privacy-policy.component.html',
-  styleUrl: './privacy-policy.component.css'
+    standalone: true,
+    selector: 'app-privacy-policy',
+    imports: [],
+    templateUrl: './privacy-policy.component.html',
+    styleUrl: './privacy-policy.component.css'
 })
 export class PrivacyPolicyComponent {
 
