@@ -1,11 +1,11 @@
 import { Component } from '@angular/core';
 
 @Component({
-  selector: 'app-req-validation',
-  standalone: true,
-  imports: [],
-  templateUrl: './req-validation.component.html',
-  styleUrl: './req-validation.component.css'
+    standalone: true,
+    selector: 'app-req-validation',
+    imports: [],
+    templateUrl: './req-validation.component.html',
+    styleUrl: './req-validation.component.css'
 })
 export class ReqValidationComponent {
 
