@@ -3,11 +3,11 @@ import { LoginComponent } from '@components/login/login.component';
 import { ProductsComponent } from '@components/products/products.component';
 
 @Component({
-  selector: 'app-home',
-  standalone: true,
-  imports: [LoginComponent, ProductsComponent],
-  templateUrl: './home.component.html',
-  styleUrl: './home.component.css'
+    standalone: true,
+    selector: 'app-home',
+    imports: [LoginComponent, ProductsComponent],
+    templateUrl: './home.component.html',
+    styleUrl: './home.component.css'
 })
 export class HomeComponent implements OnInit {
 
