@@ -1,11 +1,11 @@
 import { Component } from '@angular/core';
 
 @Component({
-  selector: 'app-not-found',
-  standalone: true,
-  imports: [],
-  templateUrl: './not-found.component.html',
-  styleUrl: './not-found.component.css'
+    standalone: true,
+    selector: 'app-not-found',
+    imports: [],
+    templateUrl: './not-found.component.html',
+    styleUrl: './not-found.component.css'
 })
 export class NotFoundComponent {
 
